+++ src/App.js
@@ -1,6 +1,5 @@
 import './App.css';
-import Calculator from "./components/Calculator";
-
+import Calculator from './components/Calculator';
 
 function App() {
   return (
