+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import Calculator from "./components/Calculator.js";
+import Calculator from "./components/Calculator";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
